create product proposition ui page
