+++ CLIENT/CLIENT.ProcureAccess/src/app/features/filters/pages/filters-container/filters-container.scss
@@ -4,32 +4,3 @@
     height: calc(100vh - 260px);
     overflow: auto;
 }
-
-.accordion-action-buttons {
-    display: flex;
-    justify-content: end;
-}
-
-.filters-accordion {
-    margin-block: 8px;
-}
-
-.list-selection-container {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-}
-
-.example-form {
-  min-width: 150px;
-  max-width: 500px;
-  width: 100%;
-}
-
-.example-full-width {
-  width: 100%;
-}
-
-.example-chip-list {
-  width: 100%;
-}
