+++ CLIENT/CLIENT.ProcureAccess/src/app/features/filters/selection/filters-selection.html
@@ -1,0 +1,72 @@
+<div class="accordion-action-buttons">
+  <button matButton (click)="accordion().openAll()">Expand All</button>
+  <button matButton (click)="accordion().closeAll()">Collapse All</button>
+</div>
+
+<mat-accordion class="filters-accordion" multi>
+    @for (filterType of store.filterTypes(); track filterType; let idx = $index, e = $even)  {
+        <mat-expansion-panel>
+            <mat-expansion-panel-header>
+                <mat-panel-title>{{ filterType.name }}</mat-panel-title>
+                <mat-panel-description>
+                    {{ productTypes().length }} selected
+                </mat-panel-description>
+            </mat-expansion-panel-header>
+
+            <p>{{ filterType.description }}</p>
+            @if (filterType.name == EnFilterType.productType) {
+                <form> <!-- [formGroup]="formGroup1()"-->
+                    <mat-form-field class="example-chip-list">
+                        <mat-label>{{ filterType.name }}</mat-label>
+                        <input
+                            name="currentProductType"
+                            placeholder=""
+                            #productTypeInput
+                            matInput
+                            [formControl]="myControl"
+                            [matChipInputFor]="chipGrid"
+                            [matAutocomplete]="auto"
+                            [matChipInputSeparatorKeyCodes]="separatorKeysCodes"
+                            (matChipInputTokenEnd)="add($event)"
+                        />
+                        <!-- [(ngModel)]="currentProductType" -->
+                        <mat-autocomplete #auto="matAutocomplete" (optionSelected)="selected($event); productTypeInput.value = ''">
+                            @for (productType of filteredProductTypes | async; track productType) {
+                                <mat-option [value]="productType">{{ productType }}</mat-option>
+                            }
+                        </mat-autocomplete>
+                    </mat-form-field>
+                </form>
+                <mat-chip-grid #chipGrid aria-label="Product Type selection">
+                    @for (productType of productTypes(); track $index) {
+                        <mat-chip-row (removed)="remove(productType)">
+                            {{ productType }}
+                            <button matChipRemove [attr.aria-label]="'remove ' + productType">
+                                <mat-icon>cancel</mat-icon>
+                            </button>
+                        </mat-chip-row>
+                    }
+                </mat-chip-grid>
+            } @else {
+                <div class="list-selection-container">
+                    <div>
+                        <mat-selection-list #filterTypes1>
+                            @for (criteriaFilter of filterType.criteriaFilters.slice(0, (filterType.criteriaFilters.length + 1) / 2); track criteriaFilter) {
+                                <mat-list-option togglePosition="before" (selectedChange)="changeFilterTypeSelection($event, criteriaFilter.id)">{{ criteriaFilter.name }}</mat-list-option>
+                            }
+                        </mat-selection-list>
+                    </div>
+                    <div>
+                        <mat-selection-list #filterTypes2>
+                            @for (criteriaFilter of filterType.criteriaFilters.slice((filterType.criteriaFilters.length + 1) / 2, filterType.criteriaFilters.length); track criteriaFilter) {
+                                <mat-list-option togglePosition="before" (selectedChange)="changeFilterTypeSelection($event, criteriaFilter.id)">{{ criteriaFilter.name }}</mat-list-option>
+                            }
+                        </mat-selection-list>
+                    </div>
+                </div>
+            }
+        </mat-expansion-panel>
+    }
+</mat-accordion>
+
+<p style="margin-left: auto; text-align: right;">{{ selectedFilterTypeIds().length }} selected</p>
