+++ CLIENT/CLIENT.ProcureAccess/src/app/features/filters/selection/filters-selection.scss
@@ -1,0 +1,28 @@
+.accordion-action-buttons {
+    display: flex;
+    justify-content: end;
+}
+
+.filters-accordion {
+    margin-block: 8px;
+}
+
+.list-selection-container {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+}
+
+.example-form {
+  min-width: 150px;
+  max-width: 500px;
+  width: 100%;
+}
+
+.example-full-width {
+  width: 100%;
+}
+
+.example-chip-list {
+  width: 100%;
+}
