+++ CLIENT/CLIENT.ProcureAccess/src/app/features/filters/selection/filters-selection.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { FiltersSelection } from './filters-selection';
+
+describe('FiltersSelection', () => {
+  let component: FiltersSelection;
+  let fixture: ComponentFixture<FiltersSelection>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [FiltersSelection]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(FiltersSelection);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
