+++ CLIENT/CLIENT.ProcureAccess/src/app/features/products/list/products-list.html
@@ -6,77 +6,89 @@
         <button mat-icon-button class="products-save-btn" aria-label="Products save button" (click)="openProductsSearchDialog()"><mat-icon>add_2</mat-icon></button>
     </div>
 </div> -->
-<table 
-    mat-table
-    matSort
-    class="products-table" 
-    [dataSource]="dataSource" 
-    class="mat-elevation-z8"
-    (matSortChange)="announceSortChange($event)">
-  <!-- Position Column -->
-  <ng-container matColumnDef="position">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by position"> No. </th>
-    <td mat-cell *matCellDef="let element"> {{element.position}} </td>
-  </ng-container>
 
-  <!-- Actions Column -->
-  <ng-container matColumnDef="actions">
-    <th mat-header-cell *matHeaderCellDef> Actions </th>
-    <td mat-cell *matCellDef="let element" style="padding-left: 8px;">
-      <button
-        mat-icon-button
-        class="product-details-btn"
-        aria-label="Product details icon-button"
-        matTooltip="Details"
-        matTooltipPosition="above"
-        [routerLink]="['/product', element.id]"
-      >
-        <mat-icon>visibility</mat-icon>
-      </button>
-      @if (store.isAuthenticated()) {
+<div class="table-container">
+  <table
+      mat-table
+      matSort
+      class="products-table" 
+      [dataSource]="dataSource" 
+      class="mat-elevation-z8"
+      (matSortChange)="announceSortChange($event)">
+    <!-- Position Column -->
+    <ng-container matColumnDef="position">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by position"> No. </th>
+      <td mat-cell *matCellDef="let element"> {{element.position}} </td>
+    </ng-container>
+
+    <!-- Actions Column -->
+    <ng-container matColumnDef="actions">
+      <th mat-header-cell *matHeaderCellDef> Actions </th>
+      <td mat-cell *matCellDef="let element" style="padding-left: 8px;">
         <button
           mat-icon-button
-          class="product-save-btn"
-          aria-label="Product save icon-button"
-          matTooltip="Save to favorites" 
-          matTooltipPosition="right"
-          (click)="addProductFavorite(element.name)">
-          <mat-icon>add</mat-icon>
+          class="product-details-btn"
+          aria-label="Product details icon-button"
+          matTooltip="Details"
+          matTooltipPosition="above"
+          [routerLink]="['/product', element.id]"
+        >
+          <mat-icon>visibility</mat-icon>
         </button>
-      }
-    </td>
-  </ng-container>
+        @if (store.isAuthenticated()) {
+          <button
+            mat-icon-button
+            class="product-save-btn"
+            aria-label="Product save icon-button"
+            matTooltip="Save to favorites" 
+            matTooltipPosition="right"
+            (click)="addProductFavorite(element.name)">
+            <mat-icon>add</mat-icon>
+          </button>
+        }
+      </td>
+    </ng-container>
 
-  <!-- Id Column -->
-  <ng-container matColumnDef="id">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by id"> Id </th>
-    <td mat-cell *matCellDef="let element"> {{element.id}} </td>
-  </ng-container>
+    <!-- Id Column -->
+    <ng-container matColumnDef="id">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by id"> Id </th>
+      <td mat-cell *matCellDef="let element"> {{element.id}} </td>
+    </ng-container>
 
-  <!-- Name Column -->
-  <ng-container matColumnDef="name">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by name"> Name </th>
-    <td mat-cell *matCellDef="let element"> {{element.name}} </td>
-  </ng-container>
+    <!-- Name Column -->
+    <ng-container matColumnDef="name">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by name"> Name </th>
+      <td mat-cell *matCellDef="let element"> {{element.name}} </td>
+    </ng-container>
 
-  <!-- Link Column -->
-  <ng-container matColumnDef="link">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by link"> Link </th>
-    <td mat-cell *matCellDef="let element"> <a [href]="'https://' + element.link" target="_blank"> {{element.link}} </a></td>
-  </ng-container>
+    <!-- Link Column -->
+    <ng-container matColumnDef="link">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by link"> Link </th>
+      <td mat-cell *matCellDef="let element"> <a [href]="'https://' + element.link" target="_blank"> {{element.link}} </a></td>
+    </ng-container>
 
-  <!-- Description Column -->
-  <ng-container matColumnDef="description">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by description"> Functionality </th>
-    <td mat-cell *matCellDef="let element"> {{element.description}} </td>
-  </ng-container>
+    <!-- Description Column -->
+    <ng-container matColumnDef="description">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by description"> Functionality </th>
+      <td mat-cell *matCellDef="let element"> {{element.description}} </td>
+    </ng-container>
 
-  <!-- Type Column -->
-  <ng-container matColumnDef="type">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by type"> Type </th>
-    <td mat-cell *matCellDef="let element"> {{element.type}} </td>
-  </ng-container>
+    <!-- Type Column -->
+    <ng-container matColumnDef="type">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by type"> Type </th>
+      <td mat-cell *matCellDef="let element"> {{element.type}} </td>
+    </ng-container>
 
-  <tr mat-header-row *matHeaderRowDef="displayedColumns; sticky: true"></tr>
-  <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-</table>
+    <tr mat-header-row *matHeaderRowDef="displayedColumns; sticky: true"></tr>
+    <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
+  </table>
+</div>
+
+<div class="product-list-actions">
+  @if (store.isAuthenticated()) {
+    <button 
+      matButton="filled"
+      aria-label="Product proposition button" 
+      [routerLink]="'/product-proposition'">Propose product</button>
+  }
+</div>
