+++ CLIENT/CLIENT.ProcureAccess/src/app/features/products/pages/products-container/products-container.html
@@ -7,7 +7,7 @@
         <input matInput placeholder="Last name, First name" formControlName="firstCtrl" required>
       </mat-form-field> -->
       <!--<pa-products-filters></pa-products-filters> [formGroup1]="criteriaFiltersFormGroup"-->
-      <pa-filters-container [formGroup1]="selectedCriteriaFiltersFormGroup" (selectedFilterTypeIdsOut)="fdsa($event)"></pa-filters-container>
+      <pa-filters-container (selectedFilterTypeIdsOut)="fdsa($event)"></pa-filters-container>
       <div class="stepper-actions-container">
         <button matButton (click)="stepper.reset()">Reset</button>
         <button matButton matStepperNext (click)="asdf()">Enter</button>
