+++ CLIENT/CLIENT.ProcureAccess/src/app/features/products/proposition/product-proposition.html
@@ -1,0 +1,52 @@
+<p>Please enter information about the product.</p>
+<p>Not all fields have to be filled out. The more information you can provide, 
+    the easier it is for our admin to recheck everything.</p>
+
+<!--
+Name, Link, Description, Tests, Product Type, App Type, Parts, Contact Persons?
+-->
+<div class="container">
+  <div class="header">
+    <div class="title"><mat-icon>add_2</mat-icon> <span>Product</span></div>
+  </div>
+  <form class="product-form" #addProductForm="ngForm" (ngSubmit)="onSubmit(addProductForm, $event)">
+    <mat-form-field>
+      <mat-label>Name</mat-label>
+      <input matInput type="text" name="name" [(ngModel)]="name" required #nameInput="ngModel"
+        maxlength="50" />
+      @if (nameInput.invalid && nameInput.touched) {
+      <mat-error>Name is required.</mat-error>
+      }
+    </mat-form-field>
+    <mat-form-field id="link-field">
+      <mat-label>Link</mat-label>
+      <input matInput type="text" name="link" [(ngModel)]="link" required #linkInput="ngModel"
+        minlength="6" maxlength="200" />
+      @if (linkInput.invalid && linkInput.touched) {
+      <mat-error>Link is required.</mat-error>
+      }
+    </mat-form-field>
+    <mat-form-field id="description-field">
+      <mat-label>Description</mat-label>
+      <!-- TODO: make text area markdown input -->
+      <textarea matInput type="text" name="description" [(ngModel)]="description" required #descriptionInput="ngModel"
+        minlength="6" maxlength="2000"></textarea>
+      @if (descriptionInput.invalid && descriptionInput.touched) {
+        <mat-error>Description is required.</mat-error>
+      }
+    </mat-form-field>
+
+    <pa-filters-selection (selectedFilterTypeIdsOut)="fdsa($event)"
+    ></pa-filters-selection>
+
+    @if (formErrorMessage) {
+        <span class="error-message ng-touched ng-invalid">{{ formErrorMessage }}</span>
+    }
+    <div class="buttons">
+      <button mat-button color="primary" (click)="null" type="button">Reset</button>
+      <button id="send-button" mat-flat-button color="primary" type="submit" [disabled]="addProductForm.invalid">
+        Send
+      </button>
+    </div>
+  </form>
+</div>
